+++ src/main/java/frc/robot/Robot.java
@@ -14,7 +14,7 @@ import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import frc.robot.config.Config;
 import frc.robot.nettables.ControlCtrlNetTable;
 import frc.robot.nettables.VisionCtrlNetTable;
-import frc.robot.subsystems.DriveBase;
+import frc.robot.subsystems.DriveBaseHolder;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
@@ -50,7 +50,7 @@ public class Robot extends TimedRobot {
     public void robotInit() {
         // Instantiate our RobotContainer.  This will perform all our button bindings, and put our
         // autonomous chooser on the dashboard.
-        DriveBase.init();
+        DriveBaseHolder.init();
         
         m_robotContainer = new RobotContainer();
