+++ src/main/java/frc/robot/commands/TankDrive.java
@@ -3,8 +3,9 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import java.util.function.Supplier;
-import frc.robot.config.Config;
+
 import frc.robot.subsystems.DriveBase;
+import frc.robot.subsystems.DriveBaseHolder;
 
 /**
  * Class to allow for usage of WPI and Comand functions with TankDrive
@@ -17,6 +18,8 @@ public class TankDrive extends CommandBase {
   private final boolean squareInputs;
   private final boolean initBrake;
 
+  private final DriveBase driveBase;
+
   /**
    * 
    * @param leftSpeed The values to use for the left side
@@ -27,35 +30,34 @@ public class TankDrive extends CommandBase {
   protected TankDrive(Supplier<Double> leftSpeed, Supplier<Double> rightSpeed, boolean squareInputs, boolean initBrake) {
     //Ensure that this command is the only one to run on the drive base
     //Requires must be included to use this command as a default command for the drive base
-    addRequirements(DriveBase.getInstance());
-
+    this.driveBase = DriveBaseHolder.getInstance();
     this.leftSpeed = leftSpeed;
     this.rightSpeed = rightSpeed;
     this.squareInputs = squareInputs;
     this.initBrake = initBrake;
-
+    addRequirements(this.driveBase);
   }
 
 
   @Override
   public void initialize() {
     //Prepare for driving by human
-    DriveBase.getInstance().setOpenLoopVoltage();
-    DriveBase.getInstance().setBrakeMode(initBrake);
+    this.driveBase.setOpenLoopVoltage();
+    this.driveBase.setBrakeMode(initBrake);
   }
 
  
   @Override
   public void execute() {
     //Pass values to drive base to make the robot move
-    DriveBase.getInstance().tankDrive(leftSpeed.get(),rightSpeed.get(),squareInputs);
+    this.driveBase.tankDrive(leftSpeed.get(),rightSpeed.get(),squareInputs);
   }
 
 
   @Override
   public void end(boolean interrupted) {
     //go back to disabled mode
-    DriveBase.getInstance().setDisabledMode();
+    this.driveBase.setDisabledMode();
 
   
   }
